feat(i18n): move customer routes under app/[locale]/ and scaffold locale layout
Phase 1 task 2 of next-intl migration. Physically relocates the customer-
facing surface under an [locale] segment so next-intl's locale-prefixed
routing can take over.

What moved:
- app/page.tsx → app/[locale]/page.tsx
- app/(features)/{auth,bookings,vip}/** → app/[locale]/(features)/...
- app/{play-and-food,golf-club-rental,course-rental}/** → app/[locale]/...

What's new:
- app/[locale]/layout.tsx — owns <html lang>, NextIntlClientProvider,
  Poppins font, GTM/structured-data scripts, full SEO metadata, and
  setRequestLocale(locale). generateStaticParams returns all locales.
- app/auth/layout.tsx — minimal <html>+Providers wrapper for the OAuth
  callback error page (app/auth/error/), which still lives at root.

What changed:
- app/layout.tsx is now a passthrough returning children, with only
  metadataBase + viewport. The rich metadata moved to the locale layout.
- app/liff/layout.tsx now provides its own <html>+Providers shell, since
  the root layout no longer does.
- middleware.ts matcher temporarily extended with /(en|th|ko|ja|zh)/:path*
  so locale-prefixed paths get processed. Full middleware integration is
  task 3.

Stays at root (untouched): app/api/**, app/liff/** (content), app/auth/error/,
app/error.tsx, app/loading.tsx, app/not-found.tsx, app/globals.css,
app/providers.tsx.

Import-path fixes accompanying the move:
- VIP context import in components/vip/* and link/profile/packages pages
  updated from relative '../../app/(features)/vip/...' to alias path.
- Layout import in [locale]/{course-rental,golf-club-rental,play-and-food}
  updated to point at the moved booking Layout component.
- Two pre-existing <a> hrefs that the route move newly trips
  no-html-link-for-pages on: ConfirmationContent now uses next/link's
  Link; QuickLinks (LIFF page hopping out to /golf-club-rental
  intentionally) gets an eslint-disable-next-line.

Verified:
- npm run typecheck: clean
- npm run build: clean; static prerendering correctly emits .html with
  <!DOCTYPE html><html lang="en"|"th"|... per locale
- dev smoke test on /, /bookings, /en, /en/bookings: all 200, content
  renders. Dev mode emits a __next_root_layout_missing_tags=["html"]
  warning because the *root* app/layout.tsx no longer owns <html> — this
  is the expected/required structure for the [locale] pattern and does
  not affect production output.
