+++ app/[locale]/(features)/auth/layout.tsx
