+++ app/[locale]/(features)/bookings/components/BayInfoModal.tsx
