+++ app/[locale]/(features)/bookings/components/booking/Layout.tsx
