+++ app/[locale]/(features)/bookings/components/booking/steps/BookingDetails.tsx
