+++ app/[locale]/(features)/bookings/confirmation/page.tsx
