+++ app/[locale]/(features)/bookings/page.tsx
