+++ app/[locale]/(features)/bookings/types/index.ts
