+++ app/[locale]/(features)/vip/dashboard/page.tsx
