+++ app/[locale]/(features)/vip/packages/page.tsx
@@ -2,17 +2,16 @@
 
 import React, { useEffect } from 'react';
 import { useRouter } from 'next/navigation';
-import ProfileView from '../../../../components/vip/ProfileView'; 
-import { useVipContext } from '../contexts/VipContext'; 
-import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert'; // Needs installation
-import { AlertCircle, Loader2 } from 'lucide-react';
+import PackagesList from '@/components/vip/PackagesList'; // To be created
+import { useVipContext } from '../contexts/VipContext';
+import { Loader2 } from 'lucide-react';
 
-const VipProfilePage = () => {
+const VipPackagesPage = () => {
   const { vipStatus, isLoadingVipStatus, session } = useVipContext();
   const router = useRouter();
 
   // Redirect unlinked users to link-account page
-  // Note: linked_unmatched users can access profile to edit their VIP data
+  // Note: linked_unmatched users can access this page but will see empty state
   useEffect(() => {
     if (!isLoadingVipStatus && vipStatus && (
       vipStatus.status === 'not_linked' || 
@@ -26,7 +25,7 @@ const VipProfilePage = () => {
     return (
       <div className="flex flex-col items-center justify-center min-h-[calc(100vh-300px)]">
         <Loader2 className="h-10 w-10 animate-spin text-primary mb-4" />
-        <p className="text-muted-foreground">Loading profile information...</p>
+        <p className="text-muted-foreground">Loading package information...</p>
       </div>
     );
   }
@@ -40,22 +39,16 @@ const VipProfilePage = () => {
       </div>
     );
   }
-  
+
   return (
-    <div className="container mx-auto max-w-3xl py-8 px-4">
-       {vipStatus && vipStatus.status === 'not_linked' && (
-        <Alert variant="destructive" className="mb-6">
-          <AlertCircle className="h-4 w-4" />
-          <AlertTitle>Account Not Linked</AlertTitle>
-          <AlertDescription>
-            Your account is not linked to a VIP profile. Please link your account from the dashboard to access VIP features.
-            The profile information below might be limited.
-          </AlertDescription>
-        </Alert>
-      )}
-      <ProfileView />
+    <div className="space-y-8">
+      <div>
+        <h1 className="text-2xl font-bold text-gray-900">My Packages</h1>
+        <p className="text-muted-foreground">View your active and past lesson or practice packages.</p>
+      </div>
+      <PackagesList /> {/* This component will handle actual data fetching and display */}
     </div>
   );
 };
 
-export default VipProfilePage; 
+export default VipPackagesPage;
