+++ app/[locale]/(features)/vip/profile/page.tsx
@@ -2,16 +2,17 @@
 
 import React, { useEffect } from 'react';
 import { useRouter } from 'next/navigation';
-import PackagesList from '../../../../components/vip/PackagesList'; // To be created
-import { useVipContext } from '../contexts/VipContext';
-import { Loader2 } from 'lucide-react';
+import ProfileView from '@/components/vip/ProfileView';
+import { useVipContext } from '../contexts/VipContext'; 
+import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert'; // Needs installation
+import { AlertCircle, Loader2 } from 'lucide-react';
 
-const VipPackagesPage = () => {
+const VipProfilePage = () => {
   const { vipStatus, isLoadingVipStatus, session } = useVipContext();
   const router = useRouter();
 
   // Redirect unlinked users to link-account page
-  // Note: linked_unmatched users can access this page but will see empty state
+  // Note: linked_unmatched users can access profile to edit their VIP data
   useEffect(() => {
     if (!isLoadingVipStatus && vipStatus && (
       vipStatus.status === 'not_linked' || 
@@ -25,7 +26,7 @@ const VipPackagesPage = () => {
     return (
       <div className="flex flex-col items-center justify-center min-h-[calc(100vh-300px)]">
         <Loader2 className="h-10 w-10 animate-spin text-primary mb-4" />
-        <p className="text-muted-foreground">Loading package information...</p>
+        <p className="text-muted-foreground">Loading profile information...</p>
       </div>
     );
   }
@@ -39,16 +40,22 @@ const VipPackagesPage = () => {
       </div>
     );
   }
-
+  
   return (
-    <div className="space-y-8">
-      <div>
-        <h1 className="text-2xl font-bold text-gray-900">My Packages</h1>
-        <p className="text-muted-foreground">View your active and past lesson or practice packages.</p>
-      </div>
-      <PackagesList /> {/* This component will handle actual data fetching and display */}
+    <div className="container mx-auto max-w-3xl py-8 px-4">
+       {vipStatus && vipStatus.status === 'not_linked' && (
+        <Alert variant="destructive" className="mb-6">
+          <AlertCircle className="h-4 w-4" />
+          <AlertTitle>Account Not Linked</AlertTitle>
+          <AlertDescription>
+            Your account is not linked to a VIP profile. Please link your account from the dashboard to access VIP features.
+            The profile information below might be limited.
+          </AlertDescription>
+        </Alert>
+      )}
+      <ProfileView />
     </div>
   );
 };
 
-export default VipPackagesPage; 
+export default VipProfilePage;
