+++ app/[locale]/course-rental/layout.tsx
