+++ app/[locale]/golf-club-rental/layout.tsx
