+++ app/[locale]/play-and-food/page.tsx
@@ -3,7 +3,7 @@
 import Image from 'next/image';
 import Link from 'next/link';
 import { useState } from 'react';
-import { Layout } from '@/app/(features)/bookings/components/booking/Layout';
+import { Layout } from '@/app/[locale]/(features)/bookings/components/booking/Layout';
 import { getPlayFoodPackages } from '@/types/play-food-packages';
 import { usePricingLoader } from '@/lib/pricing-hook';
